requirements.txt
pyyaml
fastapi
uvicorn
py3langid
iso-639
openai
opencv-python
gradio
httpx[socks]
astunparse
langchain
google-generativeai